Use Native Web Crypto API
**Context:**
* Crypto mainly used for broadcast of sensitive data.
* Users only have decryption keys for non-conflicting papers.

**Summary of changes:**
* Remove CryptoJS dependency at React client.
* Replace with Native Web Crypto API.
* Restructure code for async decryption.
* Added debugging mode, triggered at Flask server.
* Python library pycryptodome replaced with 

**Rationale:** 
* We were using crypto-js with ECB encryption.
* Development on crypto-js halted -- now deprecated.
* Native Web Crypto API widely supported.
* ECB not supported in Native Web Crypto.
* Now using AES-GCM: better security than ECB mode.
* Built-in authentication to catch corrupted data.
* No need to pad encrypted message to specific lengths.

The **main challenge** is that crypto-js is synchronous, whereas Native Web Crypto (the new library) is asynchronous. To adapt to the new paradigm required some refactoring of the crypto call stack at the React client:
* KeyContext: stores keys, handles decryption, supplies API
* GridContext and QueueContext: clients of that API
* In the **synchronous** (old) approach: methods in GridContext and QueueContext made calls to decryptObjectOrNull (the API for KeyContext) -- which returned a decrypted object. In a couple cases this was in the middle of a for-loop over an array, to decrypt all papers in the queue or in the grid.
* In the **async** (new) approach: GridContext and QueueContext make calls to decryptThenHandleObj or decryptThenHandleArray (the new API for KeyContext) -- which perform the decryption and then issue to result to a callback handler supplied by the calling method. This breaks the async chain, meaning the async functions are isolated to KeyContext.

A **secondary challenge** addressed by this PR is that we are now using AES-GCM instead of ECB. The main motivation is that ECB is not supported by Native Web Crypto. But also AES-GCM happens to offer better security than ECB mode. This change required also replacing the encryption approach at the Flask server.

Finally an optional **debugging** mode was added wherein the server could add an extra field to the encrypted object, like this:
 * `[ {"nid": ...2869...false}]} | eWk2CASE...3864...5xoVFhE= ]`
This field contains the beginning and end (and length in the middle) of both the unencrypted and encrypted strings. It can then be checked againast the equivalent strings at the receiving client, to make sure that the encryption matches. If they match no action is taken, but otherwise warnings are issued.